+++ src/EntityGraphQL/Compiler/EntityQuery/EntityQueryNodeVisitor.cs
@@ -154,13 +154,13 @@ namespace EntityGraphQL.Compiler.EntityQuery
                 return enumOrConstantValue;
             }
 
-            if(schemaType.IsEnum)
+            if (schemaType.IsEnum)
             {
-                return Expression.Constant(Enum.Parse(schemaType.TypeDotnet, context.GetText()));                
+                return Expression.Constant(Enum.Parse(schemaType.TypeDotnet, context.GetText()));
             }
 
             var gqlField = schemaType.GetField(field, requestContext);
-            (var exp, _) = gqlField.GetExpression(gqlField.ResolveExpression!, currentContext, null, null, new Dictionary<string, object>(), null, null, new List<GraphQLDirective>(), false, new Util.ParameterReplacer());
+            (var exp, _) = gqlField.GetExpression(gqlField.ResolveExpression!, currentContext, null, null, null, new Dictionary<string, object>(), null, null, new List<GraphQLDirective>(), false, new Util.ParameterReplacer());
             return exp!;
         }