+++ src/EntityGraphQL/Compiler/GqlNodes/CompileContext.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
+using EntityGraphQL.Schema;
 
 namespace EntityGraphQL.Compiler
 {
@@ -11,6 +12,7 @@ namespace EntityGraphQL.Compiler
     {
         private readonly HashSet<Type> servicesCollected = new();
         private readonly Dictionary<ParameterExpression, object> constantParameters = new();
+        private readonly Dictionary<IField, ParameterExpression> constantParametersForField = new();
 
         public HashSet<Type> Services { get => servicesCollected; }
         public IReadOnlyDictionary<ParameterExpression, object> ConstantParameters { get => constantParameters; }
@@ -23,9 +25,18 @@ namespace EntityGraphQL.Compiler
             }
         }
 
-        public void AddConstant(ParameterExpression parameterExpression, object value)
+        public void AddConstant(IField? fromField, ParameterExpression parameterExpression, object value)
         {
             constantParameters[parameterExpression] = value;
+            if (fromField != null)
+                constantParametersForField[fromField] = parameterExpression;
+        }
+
+        public ParameterExpression? GetConstantParameterForField(IField field)
+        {
+            if (constantParametersForField.TryGetValue(field, out var param))
+                return param;
+            return null;
         }
     }
 }