+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLListSelectionField.cs
@@ -59,9 +59,7 @@ namespace EntityGraphQL.Compiler
                 listContext = ReplaceContext(replacementNextFieldContext!, isRoot, replacer, listContext!);
                 nextFieldContext = Expression.Parameter(listContext.Type.GetEnumerableOrArrayType()!, $"{nextFieldContext.Name}2");
             }
-            (listContext, var argumentValues) = Field?.GetExpression(listContext!, replacementNextFieldContext, ParentNode!, schemaContext, ResolveArguments(Arguments), docParam, docVariables, directives, contextChanged, replacer) ?? (ListExpression, null);
-            if (argumentValues != null)
-                compileContext.AddConstant(Field!.ArgumentParam!, argumentValues);
+            (listContext, var argumentParam) = Field?.GetExpression(listContext!, replacementNextFieldContext, ParentNode!, schemaContext, compileContext, Arguments, docParam, docVariables, directives, contextChanged, replacer) ?? (ListExpression, null);
             if (listContext == null)
                 return null;
 
@@ -78,7 +76,7 @@ namespace EntityGraphQL.Compiler
                 return listContext;
             }
 
-            (listContext, selectionFields, nextFieldContext) = ProcessExtensionsSelection(listContext, selectionFields, nextFieldContext, contextChanged, replacer);
+            (listContext, selectionFields, nextFieldContext) = ProcessExtensionsSelection(listContext, selectionFields, nextFieldContext, argumentParam, contextChanged, replacer);
 
             if (HasServices)
                 compileContext.AddServices(Field!.Services);