+++ src/EntityGraphQL/Schema/FieldExtensions/IFieldExtension.cs
@@ -24,12 +24,12 @@ namespace EntityGraphQL.Schema.FieldExtensions
         /// </summary>
         /// <param name="field"></param>
         /// <param name="expression">The current expression for the field</param>
-        /// <param name="argExpression">The ParameterExpression used for accessing the arguments. Null if the field has no augments</param>
+        /// <param name="argumentParam">The ParameterExpression used for accessing the arguments. Null if the field has no augments</param>
         /// <param name="arguments">The values of the arguments. Null if field have no arguments</param>
         /// <param name="context">The context of the schema</param>
         /// <param name="servicesPass">True if this is the second visit. This means the object graph is built and we are now bringing in fields that use services</param>
         /// <returns></returns>
-        Expression? GetExpression(IField field, Expression expression, ParameterExpression? argExpression, dynamic? arguments, Expression context, IGraphQLNode? parentNode, bool servicesPass, ParameterReplacer parameterReplacer);
+        Expression? GetExpression(IField field, Expression expression, ParameterExpression? argumentParam, dynamic? arguments, Expression context, IGraphQLNode? parentNode, bool servicesPass, ParameterReplacer parameterReplacer);
         /// <summary>
         /// Called when the field is being finalized for execution but we have not yet selected the fields for selection on this expression
         /// Not called for GraphQLFieldType.Scalar
@@ -46,11 +46,13 @@ namespace EntityGraphQL.Schema.FieldExtensions
         /// 
         /// This should be thread safe
         /// </summary>
-        /// <param name="fieldType">Type of field being built. ListSelection or ObjectProjection</param>
         /// <param name="baseExpression">ListSelection: The expression used to add .Select() to. ObjectProjection: the base expression which fields are selected from</param>
         /// <param name="selectionExpressions">ListSelection: The selection fields used in .Select(). ObjectProjection: The fields used in the new { field1 = ..., field2 = ... }</param>
+        /// <param name="selectContextParam"></param>
+        /// <param name="argumentParam"></param>
+        /// <param name="parameterReplacer"></param>
         /// <returns></returns>
-        (Expression baseExpression, Dictionary<IFieldKey, CompiledField> selectionExpressions, ParameterExpression? selectContextParam) ProcessExpressionSelection(Expression baseExpression, Dictionary<IFieldKey, CompiledField> selectionExpressions, ParameterExpression? selectContextParam, bool servicesPass, ParameterReplacer parameterReplacer);
+        (Expression baseExpression, Dictionary<IFieldKey, CompiledField> selectionExpressions, ParameterExpression? selectContextParam) ProcessExpressionSelection(Expression baseExpression, Dictionary<IFieldKey, CompiledField> selectionExpressions, ParameterExpression? selectContextParam, ParameterExpression? argumentParam, bool servicesPass, ParameterReplacer parameterReplacer);
         /// <summary>
         /// Called when the field is being finalized for execution
         ///