+++ src/EntityGraphQL/Schema/FieldToResolve.cs
@@ -10,36 +10,36 @@ public class FieldToResolveWithArgs<TContext, TParams> : Field
 
     public Field Resolve(Expression<Func<TContext, TParams, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, true);
         return this;
     }
     public Field ResolveWithService<TService>(Expression<Func<TContext, TParams, TService, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, true);
         Services = new[] { typeof(TService) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TParams, TService1, TService2, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, true);
         Services = new[] { typeof(TService1), typeof(TService2) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TParams, TService1, TService2, TService3, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, true);
         Services = new[] { typeof(TService1), typeof(TService2), typeof(TService3) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, true);
         Services = new[] { typeof(TService1), typeof(TService2), typeof(TService3), typeof(TService4) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, true);
         Services = new[] { typeof(TService1), typeof(TService2), typeof(TService3), typeof(TService4), typeof(TService5) };
         return this;
     }
@@ -53,37 +53,37 @@ public class FieldToResolve<TContext> : Field
 
     public Field Resolve(Expression<Func<TContext, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, false);
+        SetUpField(fieldExpression, false, false);
         return this;
     }
 
     public Field ResolveWithService<TService>(Expression<Func<TContext, TService, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, false);
         Services = new[] { typeof(TService) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TService1, TService2, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, false);
         Services = new[] { typeof(TService1), typeof(TService2) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TService1, TService2, TService3, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, false);
         Services = new[] { typeof(TService1), typeof(TService2), typeof(TService3) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TService1, TService2, TService3, TService4, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, false);
         Services = new[] { typeof(TService1), typeof(TService2), typeof(TService3), typeof(TService4) };
         return this;
     }
     public Field ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
     {
-        SetUpField(fieldExpression, true);
+        SetUpField(fieldExpression, true, false);
         Services = new[] { typeof(TService1), typeof(TService2), typeof(TService3), typeof(TService4), typeof(TService5) };
         return this;
     }