+++ src/EntityGraphQL/Schema/MethodField.cs
@@ -58,7 +58,7 @@ namespace EntityGraphQL.Schema
                     {
                         continue;
                     }
-                    
+
                     if (!schema.HasType(inputType) && options.AutoCreateInputTypes)
                     {
                         AddInputTypesInArguments(schema, options.AutoCreateInputTypes, inputType);
@@ -191,7 +191,7 @@ namespace EntityGraphQL.Schema
             return result;
         }
 
-        public override (Expression? expression, object? argumentValues) GetExpression(Expression fieldExpression, Expression? fieldContext, IGraphQLNode? parentNode, ParameterExpression? schemaContext, Dictionary<string, object> args, ParameterExpression? docParam, object? docVariables, IEnumerable<GraphQLDirective> directives, bool contextChanged, ParameterReplacer replacer)
+        public override (Expression? expression, ParameterExpression? argumentParam) GetExpression(Expression fieldExpression, Expression? fieldContext, IGraphQLNode? parentNode, ParameterExpression? schemaContext, CompileContext? compileContext, IReadOnlyDictionary<string, object> args, ParameterExpression? docParam, object? docVariables, IEnumerable<GraphQLDirective> directives, bool contextChanged, ParameterReplacer replacer)
         {
             var result = fieldExpression;