+++ src/EntityGraphQL/Schema/SchemaBuilder.cs
@@ -204,7 +204,7 @@ namespace EntityGraphQL.Schema
                 // see if there is a direct type mapping from the expression return to to something.
                 // otherwise build the type info
                 var returnTypeInfo = schema.GetCustomTypeMapping(le.ReturnType) ?? new GqlTypeInfo(() => schema.GetSchemaType(baseReturnType, null), le.Body.Type, prop.IsNullable());
-                var field = new Field(schema, fromType, schema.SchemaFieldNamer(prop.Name), le, description, returnTypeInfo, requiredClaims);
+                var field = new Field(schema, fromType, schema.SchemaFieldNamer(prop.Name), le, description, null, returnTypeInfo, requiredClaims);
 
                 if (options.AutoCreateFieldWithIdArguments && (!schema.HasType(prop.DeclaringType!) || schema.GetSchemaType(prop.DeclaringType!, null).GqlType != GqlTypeEnum.Input))
                 {